% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp Functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{predictResponse}) and Cpp functions (\code{gibbsC} and \code{predictResponseCpp}).
}
\examples{
\dontrun{
data(illustration_data)
attach(illustration_data)
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(100, 10),
  rnC = gibbsC(100, 10)
)
print(summary(tm1)[, c(1, 3, 5, 6)])

tm2 <- microbenchmark::microbenchmark(
  vR = predictResponse(age, female, ily),
  vC = predictResponseCpp(age, female, ily)
)
print(summary(tm2)[, c(1, 3, 5, 6)])
}
}
