% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{convex_relaxation}
\alias{convex_relaxation}
\title{Convex Relaxation Function}
\usage{
convex_relaxation(
  features_mat,
  rewards_mat,
  learning_rate = 0.1,
  max_iter = 10000
)
}
\arguments{
\item{features_mat}{A matrix representing the features of the dataset.}

\item{rewards_mat}{A matrix representing the rewards associated with different actions.}

\item{learning_rate}{The learning rate for parameter updates (default is 0.1).}

\item{max_iter}{The maximum number of iterations for updating parameters (default is 10000).}
}
\value{
A matrix representing the learned parameters theta.
}
\description{
This function implements the convex relaxation algorithm for policy learning.
It iteratively updates the parameter matrix theta based on the features and rewards matrices.
}
\examples{
\dontrun{
set.seed(123)
features_mat <- matrix(rnorm(250), ncol = 5)
beta <- matrix(rnorm(10), ncol = 2)
rewards_mat <- features_mat \%*\% beta + matrix(runif(100), ncol = 2)
theta_result_convex <- convex_relaxation(features_mat, rewards_mat)
print("Theta Result (Convex Relaxation):")
print(theta_result_convex)
}

}
