% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{customGibbsR}
\alias{customGibbsR}
\title{A Custom Gibbs Sampler using R}
\usage{
customGibbsR(num_samples, thinning_factor)
}
\arguments{
\item{num_samples}{the number of samples to generate}

\item{thinning_factor}{the number of between-sample random numbers}
}
\value{
a random sample matrix of size \code{num_samples}
}
\description{
A custom implementation of Gibbs sampler using R
}
\examples{
\dontrun{
custom_samples <- customGibbsR(100, 10)
par(mfrow=c(2,1))
plot(custom_samples[, 1], type='l', main='Sampled Values for X')
plot(custom_samples[, 2], type='l', main='Sampled Values for Y')
}
}
