% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{illustration_data}
\alias{illustration_data}
\title{An Illustration Dataset}
\description{
A dataset used to illustrate the performance of \code{vaccR} and \code{vaccC}.
}
\examples{
\dontrun{
data(illustration_data)
attach(illustration_data)
tm <- microbenchmark::microbenchmark(
  vR = predictResponse(age, female, ily),
  vC = predictResponseCpp(age, female, ily)
)
print(summary(tm)[, c(1, 3, 5, 6)])
}
}
