% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{predictResponse}
\alias{predictResponse}
\title{Predict Response Using Three Inputs in R}
\usage{
predictResponse(age, is_female, has_ily)
}
\arguments{
\item{age}{the first predictor (numeric)}

\item{is_female}{the second predictor (logical)}

\item{has_ily}{the third predictor (logical)}
}
\value{
a random sample of size \code{n}
}
\description{
This function predicts a response using three input variables.
  The prediction model is described in http://www.babelgraph.org/wp/?p=358.
}
\examples{
\dontrun{
data(data)
attach(data)
prediction <- predictResponse(age, is_female, has_ily)
}
}
