\name{SA23204180-package}
\alias{SA23204180-package}
\alias{SA23204180}
\docType{package}
\title{
  Final functions for 'Statistical Computing' course
}
\description{
  Several R and Rcpp function are included in this package for illustration.
}
\author{
Yin pengcheng 
Maintainer: Yin pengcheng <yinpc@mail.ustc.edu.cn>
}

\keyword{ package }

